Refactor name of home view to landing view
